requirements.txt
pandas
openpyxl
schedule
google-ads
PyYAML
smtplib
